fix: get tweet text from config
