+++ main.py
@@ -4,6 +4,9 @@ from openai import OpenAI
 import requests
 from datetime import datetime
 from generate_prompt import generate_preset_prompt
+from config import load_config_from_yaml
+from typing import Optional, Dict, Any
+
 
 def setup_twitter_clients():
     # Twitter API v1認証（メディアアップロード用）
@@ -26,6 +29,22 @@ def setup_twitter_clients():
     )
 
     return api_v1, client_v2
+
+
+def get_tweet_text(preset_name: Optional[str] = None, config_path: Optional[str] = "prompt_config.yaml") -> str:
+    presets = load_config_from_yaml(config_path)
+
+    # 環境変数からプリセット名を取得（指定がなければデフォルト）
+    if preset_name is None:
+        preset_name = os.environ.get('PROMPT_PRESET', 'default')
+
+    # デフォルトプリセットが存在しない場合は空の辞書を使用
+    if preset_name not in presets:
+        preset_config = {}
+    else:
+        preset_config = presets[preset_name]
+
+    return presets[preset_name]["tweet_text"]
 
 
 def generate_and_post_image(prompt, tweet_text):
@@ -81,5 +100,5 @@ if __name__ == "__main__":
 
     # 画像生成と投稿
     prompt = generate_preset_prompt()
-    tweet_text = "Generated By #DALLE3 #AI #AIart #AIイラスト #銀髪"
+    tweet_text = get_tweet_text()
     generate_and_post_image(prompt, tweet_text)
