+++ prompt_config.yaml
@@ -12,6 +12,7 @@
 #   clothing: "Spring-like floral dress, focusing on flower embroidery and frill details"
 #   composition: "Capturing the full body of a child in a flower field while also expressing the surrounding nature in detail"
 #   scene: "Relaxed flower field under spring sunshine, with each surrounding flower carefully depicted"
+#   tweet_text: tweet text. include hash tag, text, etc
 
 silver:
   art_style: "The illustration is done in a flat-colored anime cel-shaded style"
@@ -23,3 +24,4 @@ silver:
   clothing: "Spring-like floral dress primarily in light pastel colors, with attention to flower embroidery and frill details"
   composition: "Focusing on a woman sitting and reading a book, while also showing the surrounding nature"
   scene: "A table and chair next to a log cabin, sitting on the chair reading a book"
+  tweet_text: "Generated By #DALLE3 #AI #AIart #AIイラスト #銀髪"
